refactor: 컴포넌트 tsx 파일 이름 컴포넌트 명으로 변경
